Move archiver service into services/repository/
