+++ services/cron/tasks_basic.go
@@ -15,7 +15,8 @@ import (
 	"code.gitea.io/gitea/services/auth"
 	"code.gitea.io/gitea/services/migrations"
 	mirror_service "code.gitea.io/gitea/services/mirror"
-	repository_service "code.gitea.io/gitea/services/repository"
+	repo_service "code.gitea.io/gitea/services/repository"
+	archiver_service "code.gitea.io/gitea/services/repository/archiver"
 )
 
 func registerUpdateMirrorTask() {
@@ -56,7 +57,7 @@ func registerRepoHealthCheck() {
 		Args:    []string{},
 	}, func(ctx context.Context, _ *user_model.User, config Config) error {
 		rhcConfig := config.(*RepoHealthCheckConfig)
-		return repository_service.GitFsck(ctx, rhcConfig.Timeout, rhcConfig.Args)
+		return repo_service.GitFsck(ctx, rhcConfig.Timeout, rhcConfig.Args)
 	})
 }
 
@@ -80,7 +81,7 @@ func registerArchiveCleanup() {
 		OlderThan: 24 * time.Hour,
 	}, func(ctx context.Context, _ *user_model.User, config Config) error {
 		acConfig := config.(*OlderThanConfig)
-		return repository_service.DeleteOldRepositoryArchives(ctx, acConfig.OlderThan)
+		return archiver_service.DeleteOldRepositoryArchives(ctx, acConfig.OlderThan)
 	})
 }
